online phase simulation merged

+++ direct_com/src/Abstracts/Node.java
@@ -5,6 +5,6 @@ package Abstracts;
  */
 public class Node{
 
-    public int dataSize;
+    public long dataSize;
 
 }
